[refactor]: 컴포넌트 별로 scss 파일 분리 (#4)
