+++ src/components/Header/HeaderNav.tsx
@@ -1,4 +1,4 @@
-import styles from './index.module.scss';
+import styles from './HeaderNav.module.scss';
 
 const HeaderNav = () => {
   return (
