+++ src/components/Header/InnerHeader.module.scss
@@ -1,0 +1,7 @@
+.wrapper_header {
+  display: flex;
+  position: relative;
+  margin: 0 auto;
+  width: 1280px;
+  height: 80px;
+}
