+++ src/components/Header/Logo.module.scss
@@ -1,0 +1,5 @@
+.txt_logo {
+  font-size: 27px;
+  font-weight: bold;
+  padding: 25px 50px 25px 0;
+}
