+++ src/components/Header/Logo.tsx
@@ -1,4 +1,4 @@
-import styles from './index.module.scss';
+import styles from './Logo.module.scss';
 
 const Logo = () => {
   return (
