+++ src/components/Header/index.module.scss
@@ -1,128 +1,4 @@
-@import '../../styles/mixins.module';
-
 .area_header {
   margin: 0;
   padding: 0;
-}
-
-.wrapper_header {
-  display: flex;
-  position: relative;
-  margin: 0 auto;
-  width: 1280px;
-  height: 80px;
-}
-
-// 타이틀
-
-.txt_logo {
-  font-size: 27px;
-  font-weight: bold;
-  padding: 25px 50px 25px 0;
-}
-
-// 네비게이션바
-
-.wrapper_nav {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  .list {
-    display: flex;
-    font-size: 16px;
-    padding: 20px 20px 20px 0;
-  }
-
-  .list_item {
-    padding-right: 10px;
-    font-weight: bold;
-  }
-
-  .link {
-    padding: 11px 10px 7px;
-  }
-}
-
-// 카테고리
-
-.wrapper_category {
-  display: flex;
-  align-items: center;
-  padding: 20px 0;
-
-  &::before {
-    width: 1px;
-    height: 14px;
-    margin-right: 20px;
-    background-color: #ddd;
-    content: '';
-  }
-
-  .btn {
-    display: inline-flex;
-    height: 40px;
-    padding: 7px 10px 0;
-    font-weight: 700;
-    font-size: 16px;
-    line-height: 27px;
-    color: #4684e9;
-  }
-
-  .ico {
-    @include header-ico(-160px, -80px, 30px, 30px);
-
-    margin: -2px 8px 0 0;
-  }
-}
-
-// bundle_util
-
-.wrapper_bundle_util {
-  margin-left: auto;
-  display: flex;
-  padding: 20px 0;
-
-  .link_search {
-    display: inline-block;
-    vertical-align: top;
-  }
-
-  .wrapper_ico {
-    display: flex;
-    align-items: center;
-  }
-
-  .ico_search {
-    @include header-ico(-60px, -80px, 24px, 24px);
-
-    margin: 8px;
-  }
-
-  .link_cart {
-    position: relative;
-    margin-left: 14px;
-  }
-
-  .ico_cart {
-    @include header-ico(-30px, -80px, 24px, 24px);
-
-    margin: 8px;
-  }
-
-  .wrapper_login {
-    display: flex;
-    align-items: center;
-    position: relative;
-    padding-left: 22px;
-  }
-
-  .link_login {
-    display: inline-flex;
-    align-items: center;
-    position: relative;
-    height: 38px;
-    padding-top: 2px;
-    font-size: 14px;
-  }
 }
